requirements.txt
discord==2.3.2
otto_api42==0.0.6
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
